requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
pytest==6.0.1
pytest-asyncio==0.14.0
PyYAML==5.3.1
